+++ CotEditor/Sources/EditorCounter.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2023 1024jp
+//  © 2014-2024 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -33,9 +33,9 @@ struct EditorInfoTypes: OptionSet {
     static let location   = Self(rawValue: 1 << 3)
     static let line       = Self(rawValue: 1 << 4)
     static let column     = Self(rawValue: 1 << 5)
-    static let unicode    = Self(rawValue: 1 << 6)
+    static let character  = Self(rawValue: 1 << 6)
     
-    static let all: Self = [.characters, .lines, .words, .location, .line, .column, .unicode]
+    static let all: Self = [.characters, .lines, .words, .location, .line, .column, .character]
     
     static let cursors: Self = [.location, .line, .column]
 }
@@ -57,7 +57,7 @@ struct EditorCountResult: Equatable {
     var line: Int?   // current line
     var column: Int?   // cursor location from the beginning of line
     
-    var unicode: String?  // Unicode of selected single character (or surrogate-pair)
+    var character: Character?  // Selected character (only when selection is single character)
 }
 
 
@@ -159,9 +159,9 @@ final actor EditorCounter {
             result.column = self.string.columnNumber(at: self.selectedRanges[0].lowerBound)
         }
         
-        if self.requiredInfo.contains(.unicode) {
-            result.unicode = (selectedStrings[0].compareCount(with: 1) == .equal)
-                ? selectedStrings[0].first?.unicodeScalars.map(\.codePoint).formatted(.list(type: .and, width: .narrow))
+        if self.requiredInfo.contains(.character) {
+            result.character = (selectedStrings[0].compareCount(with: 1) == .equal)
+                ? selectedStrings[0].first
                 : nil
         }